(1차 prototype)grounded-query/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
sentence-transformers==3.3.1
chromadb==0.5.3
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.9
